feat:search페이지 폴더 나누기
